Implement token authentication request guard
Lots of changes in this one:

- update DB migrations to use types Diesel recognizes as a timestamp
- set up a static DB connection pool that various threads can use on demand
- add auth_token table to the DB and its queryable representation
- add a bunch of dependent crates necessary to make this all happen

The benefit is that we now have two new guards we can use in our views:
db::DB, with which we can get a database connection from the pool for
local use, and auth::token::TokenAuth, which ensures that a valid
database token was presented in the `Authorization` header. The
token has to follow the same rules as those of DRF's TokenAuthentication.